% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdq-tracker-donations.R
\name{assemble_donations}
\alias{assemble_donations}
\title{Aggregate all donation data from the tracker in one file}
\usage{
assemble_donations(events = NULL, cache = TRUE)
}
\arguments{
\item{events}{\verb{[NULL]}: Optional vector of file paths to individual \code{.rds} files.}

\item{cache}{\verb{[TRUE]}: Save aggregated dataset at \code{"data/all_donations.rds"}.}
}
\value{
A \link[tibble:tibble]{tibble}.
}
\description{
Aggregate all donation data from the tracker in one file
}
\examples{
\dontrun{
assemble_donations()
}
}
