% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdq-tracker-runs.R
\name{assemble_runs}
\alias{assemble_runs}
\title{Update all runs from GDQ tracker}
\usage{
assemble_runs(events = NULL, cache = TRUE)
}
\arguments{
\item{events}{\verb{[NULL]} Optional vector of file paths to individual \code{.rds} files.}

\item{cache}{\verb{[TRUE]} Save aggregated dataset at \code{"data/all_donations.rds"}.}
}
\value{
A \link[tibble:tibble]{tibble}.
}
\description{
Update all runs from GDQ tracker
}
\examples{
\dontrun{
assemble_runs(cache = FALSE)
}
}
