% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdq-tracker-donations.R
\name{augment_donations}
\alias{augment_donations}
\title{Augment GDQ donation data with additional variables}
\usage{
augment_donations(donations)
}
\arguments{
\item{donations}{Donations \code{tibble} as returned by \code{\link[=assemble_donations]{assemble_donations()}}.}
}
\value{
A \code{\link[tibble:tibble]{tibble}}
}
\description{
Augment GDQ donation data with additional variables
}
\examples{
\dontrun{
assemble_donations() |>
  augment_donations()
}
}
