% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{event_index}
\alias{event_index}
\title{GDQ Event Index}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 22 rows and 8 columns.
}
\usage{
event_index
}
\description{
GDQ Event Index
}
\examples{
event_index
}
\keyword{datasets}
