% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gdq-data}
\alias{gdq-data}
\alias{event_index}
\alias{all_donations}
\alias{all_runs}
\title{GDQ Event Index}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 22 rows and 8 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 599331 rows and 15 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2770 rows and 11 columns.
}
\usage{
event_index

all_donations

all_runs
}
\description{
GDQ Event Index

Aggregated donation data

Aggregated run data
}
\note{
Needs manual updating with each new event
}
\keyword{datasets}
