% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{gdq-plotting}
\alias{gdq-plotting}
\alias{theme_gdq}
\alias{scale_x_year}
\alias{scale_x_year_discrete}
\alias{scale_y_currency}
\alias{scale_y_dollar}
\alias{scale_x_dollar}
\alias{scale_colorfill_gdq}
\title{Plotting helpers}
\usage{
theme_gdq(legend.position = "top", ...)

scale_x_year(...)

scale_x_year_discrete(...)

scale_y_currency(...)

scale_y_dollar(...)

scale_x_dollar(...)

scale_colorfill_gdq(...)
}
\arguments{
\item{legend.position}{\verb{["top"]}: See \code{\link[ggplot2:theme]{ggplot2::theme()}}}

\item{...}{passed to \code{\link[ggplot2:theme]{ggplot2::theme}} or other components.}
}
\description{
Plotting helpers
}
