% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gdq_donations}
\alias{gdq_donations}
\title{Aggregated donation data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 599331 rows and 15 columns.
}
\usage{
gdq_donations
}
\description{
All the GDQ donations;
}
\keyword{datasets}
