% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdq-tracker-donations.R
\name{get_donations}
\alias{get_donations}
\title{Get all donations for an event from the tracker}
\usage{
get_donations(event = "latest")
}
\arguments{
\item{event}{\verb{["latest"]} Event such as \code{"AGDQ2021"}, case insensitive. The default,
\code{"latest"}, is an alias for \verb{[latest_event()]}.}
}
\value{
A \link[tibble:tibble]{tibble}.
}
\description{
Get all donations for an event from the tracker
}
\examples{
\dontrun{

donations_sgdq2025 <- get_donations(event = "sgdq2025")
}
}
