% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdq-tracker-runs.R
\name{get_runs}
\alias{get_runs}
\title{Get runs from GDQ tracker}
\usage{
get_runs(event)
}
\arguments{
\item{event}{Event such as \code{"AGDQ2021"}, case insensitive.}
}
\value{
A \link[tibble:tibble]{tibble}.
}
\description{
Get runs from GDQ tracker
}
\examples{
\dontrun{
get_runs(event = "sgdq2021") \%>\% View()
}
}
