% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotting}
\alias{plotting}
\alias{theme_gdq}
\alias{euro_scale}
\alias{euro_axis}
\alias{scale_x_year}
\alias{scale_x_year_discrete}
\alias{scale_y_currency}
\alias{scale_y_dollar}
\alias{scale_x_dollar}
\alias{scale_colorfill_gdq}
\title{Plotting helpers}
\usage{
theme_gdq(grid = "Yy", legend.position = "top", ...)

euro_scale(x)

euro_axis(...)

scale_x_year(...)

scale_x_year_discrete(...)

scale_y_currency(...)

scale_y_dollar(...)

scale_x_dollar(...)

scale_colorfill_gdq(...)
}
\description{
Plotting helpers
}
