% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdq-tracker-runs.R
\name{summarize_runs}
\alias{summarize_runs}
\title{Summarize GDQ tracker run data by event}
\usage{
summarize_runs(runs)
}
\arguments{
\item{runs}{Run tibble as returned by \code{\link[=assemble_runs]{assemble_runs()}}.}
}
\value{
A \link[tibble:tibble]{tibble}.
}
\description{
Summarize GDQ tracker run data by event
}
\examples{
\dontrun{
assemble_runs() |>
  summarize_runs()
}
}
