% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdq-tracker.R
\name{tracker_run_index}
\alias{tracker_run_index}
\title{Get GDQ tracker event overview}
\usage{
tracker_run_index()
}
\value{
A tibble
}
\description{
Get GDQ tracker event overview
}
\examples{
\dontrun{
tracker_run_index()
}
}
