% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdq-tracker-donations.R
\name{update_tracker_donations}
\alias{update_tracker_donations}
\title{Update donation data from tracker}
\usage{
update_tracker_donations(events, ignore_cache = FALSE, in_progress = FALSE)
}
\arguments{
\item{ignore_cache}{\verb{[FALSE]}: If \code{TRUE}, ignore cached file and re-retrieve data.}

\item{in_progress}{\verb{[FALSE]}: If \code{TRUE}, donations for in-progress events are retrieved.}
}
\value{
Invisibly: Tibble of donations.
}
\description{
Update donation data from tracker
}
\examples{
\dontrun{
update_tracker_donations(events = c("agdq2021", "sgdq2021"), ignore_cache = TRUE)
}
}
