% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdq-tracker-runs.R
\name{update_tracker_runs}
\alias{update_tracker_runs}
\title{Update run data from GDQ tracker}
\usage{
update_tracker_runs(events, ignore_cache = FALSE, in_progress = FALSE)
}
\arguments{
\item{events}{Events such as \code{"agdq2019"}, case insensitive.}

\item{ignore_cache}{\verb{[FALSE]}: If \code{TRUE}, ignore cached file and re-retrieve data.}

\item{in_progress}{\verb{[FALSE]}: If \code{TRUE}, donations for in-progress events are retrieved.}
}
\value{
Nothing
}
\description{
Update run data from GDQ tracker
}
\examples{
\dontrun{
update_tracker_runs(
  events = c("agdq2021", "sgdq2021"),
  ignore_cache = TRUE,
  in_progress = TRUE
)
}
}
