% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gdq_runs}
\alias{gdq_runs}
\title{Aggregated run data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2926 rows and 11 columns.
}
\source{
\code{\link[=assemble_runs]{assemble_runs()}}
}
\usage{
gdq_runs
}
\description{
All the GDQ runs.
}
\keyword{datasets}
