% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdq-tracker-donations.R
\name{get_donation_page}
\alias{get_donation_page}
\title{Get a single page from the tracker}
\usage{
get_donation_page(event = "latest", page = 1)
}
\arguments{
\item{event}{\verb{["latest"]} Event such as \code{"AGDQ2021"}, case insensitive. The default,
\code{"latest"}, is an alias for \verb{[latest_event()]}.}

\item{page}{\verb{[1]}: Page to get, can also be a range of panges, e.g. \code{1:5}.}
}
\value{
A \link[tibble:tibble]{tibble}.
}
\description{
Get a single page from the tracker
}
\examples{
\dontrun{
get_donation_page(event = "sgdq2011", page = 1:3)
}
}
