% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdq-tracker-donations.R
\name{get_donation_page}
\alias{get_donation_page}
\title{Get a single page from the tracker}
\usage{
get_donation_page(event = "sgdq2021", page = 1)
}
\arguments{
\item{event}{Event such as \code{"AGDQ2021"}, case insensitive.}

\item{page}{Page to get, a single number.}
}
\value{
A \link[tibble:tibble]{tibble}.
}
\description{
Get a single page from the tracker
}
\examples{
\dontrun{
get_donation_page(event = "sgdq2021", page = 1)
}
}
