% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdq-tracker-donations.R
\name{get_page_count}
\alias{get_page_count}
\title{Get page count of GDQ tracker}
\usage{
get_page_count(event = "latest")
}
\arguments{
\item{event}{\verb{["latest"]} Event such as \code{"AGDQ2021"}, case insensitive. The default,
\code{"latest"}, is an alias for \code{event_index$event[nrow(event_index)]}.}
}
\value{
Page count as a \code{numeric(1)}.
}
\description{
Get page count of GDQ tracker
}
\examples{
get_page_count(event = "latest")
}
