% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdq-tracker-donations.R
\name{get_page_count}
\alias{get_page_count}
\title{Get page count of GDQ tracker}
\usage{
get_page_count(event = "sgdq2021")
}
\arguments{
\item{event}{Event such as \code{"AGDQ2021"}, case insensitive.}
}
\value{
Page count as a \code{numeric(1)}.
}
\description{
Get page count of GDQ tracker
}
\examples{
get_page_count(event = "sgdq2021")
}
