% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pipe.R
\name{\%--\%}
\alias{\%--\%}
\title{Interval operator}
\usage{
lhs \%--\% rhs
}
\arguments{
\item{start, end}{Interval start and end points.}
}
\value{
The result of calling \code{interval(start = NULL, end = NULL, tzone = tz(start))}.
}
\description{
See \code{lubridate::\link[lubridate::interval]{\%--\%}} for details.
}
\keyword{internal}
