% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdq-tracker-donations.R
\name{update_tracker_donations}
\alias{update_tracker_donations}
\title{Update donation data from GDQ tracker}
\usage{
update_tracker_donations(
  events,
  ignore_cache = FALSE,
  in_progress = FALSE,
  sound = TRUE
)
}
\arguments{
\item{events}{Events such as \code{"agdq2019"}, case insensitive.}

\item{ignore_cache}{\verb{[FALSE]}: If \code{TRUE}, ignore cached file and re-retrieve data.}

\item{in_progress}{\verb{[FALSE]}: If \code{TRUE}, donations for in-progress events are retrieved.}

\item{sound}{\verb{[TRUE]}: If \code{TRUE}, \code{beepr::beep(2)} is played after each event's
donations have been retrieved. Since scraping \emph{will} take a long time for full
events, this seemed like a good idea.}
}
\value{
Invisibly: A \link[tibble:tibble]{tibble}.
}
\description{
Update donation data from GDQ tracker
}
\examples{
\dontrun{
update_tracker_donations(
  events = c("agdq2021", "sgdq2021"),
  ignore_cache = TRUE,
  in_progress = TRUE
)
}
}
